refactor: move route test into separate files
